Improvement/data transfer

+++ src/main/java/de/hpi/msc/jschneider/utility/dataTransfer/DataSink.java
@@ -2,7 +2,7 @@ package de.hpi.msc.jschneider.utility.dataTransfer;
 
 public interface DataSink
 {
-    void write(float[] part);
+    void write(byte[] part);
 
     void close();
 }
